--- a/tools/make-umd-bundle.js
+++ b/tools/make-umd-bundle.js
@@ -1,8 +1,21 @@
-var browserify = require('browserify');
+var rollup = require('rollup');
 var fs = require('fs');
 
-var b = browserify('dist/cjs/Rx.js', {
-  baseDir: 'dist/cjs',
-  standalone: 'Rx'
+var babel = require('babel-core');
+
+rollup.rollup({
+  entry: 'dist/es6/Rx.js'
+}).then(function (bundle) {
+  var result = bundle.generate({
+    format: 'es'
+  });
+
+  var out = babel.transform(result.code, {
+    compact: false,
+    presets: [
+      ['es2015', {loose: true}]
+    ],
+  });
+
+  fs.writeFileSync('dist/global/Rx.js', out.code);
 });
-b.bundle().pipe(fs.createWriteStream('dist/global/Rx.js'));