This PR changes how the "Global" (aka umd) build is bundled and transpiled.

Before: TS-to-ES5 -> Browserify
After: TS-to-ES6 -> Rollup -> Babel-to-ES5

The new build is considerably smaller and loads much faster:


| File    | Size (before) | Size (after) | Load (before) | Load (after) |
| ------- |:-------------:|:------------:|:-------------:|:------------:|
| Rx.js         | 608K | 514K | 110ms | 34ms |
| Rx.min.js     | 222K | 168K | 95ms  | 26ms |
| Rx.min.js.map | 448K | 338K | n/a   | n/a |

To test load time, I ran `for i in {1..5}; do node -p "s=Date.now();require('path/to/bundle');Date.now()-s"; done`.

A few notes:

* The improvements come from debuping the helper code (class creation, etc.), and from rollup's single body bundling.
* In order to dedupe the helper code, tranpiling has be done post-bundling, which is why I'm using babel. AFAIK, TS can't transpile ES6 to ES5 (or can it?).
* Babel's loose-mode seemed to more closely match TS's ES5 output.
* I tested the resulting bundle by using it [nuclide](https://github.com/facebook/nuclide).