@kwonoj 2016-10-24
While it brings additional steps of generating build, I'm 👍  since outcome looks good.

One question - does sourcemap for min version points original code correctly?