@zertosh 2016-10-24
> One question - does sourcemap for min version points original code correctly?

@kwonoj The sourcemaps are generated by GCC and they're based off the post-transpiled/post-bundled output (you have to pass `debug: true` to browserify for it to preserve and output source maps). That didn't change, so the short answer is "they should be the same as before".

But this brings up another point, and it's that the source maps should be based on pre-bundled output (maybe also pre-transpiled output). This would make them quite large (prob a few MB), but maybe more useful.