@kwonoj 2016-10-24
> This would make them quite large (prob a few MB), but maybe more useful. 

(My personal 5c) I'd prefer sourcemap to be more accurate even it'll take more space. Prod. code can strip sourcemap anyway since it's not inlined.