@zertosh 2016-10-24
@kwonoj sure. Would you like me to make that change here or in another PR? (Since this PR keeps the original behavior)