@kwonoj 2016-10-24
> Would you like me to make that change here or in another PR? (Since this PR keeps the original behavior)

I agree with that, but let's consider that after got agreement on this PR and merged. I'm 👍 , need additional approval to check in.