@Rich-Harris 2016-10-24
Hey – Rollup author here, I got alerted to this thread. The sourcemap thing is an issue I've had to deal with in a few different contexts, so I wrote a tool called [sorcery](https://github.com/rich-harris/sorcery) to generate a sourcemap that accounts for an arbitrary number of sourcemap-generating transformations. It's not bulletproof (nothing involving sourcemaps is, in my experience!) but it might be useful so I thought I'd mention it.

It ought to be possible to dedupe helpers using [rollup-plugin-babel](https://github.com/rollup/rollup-plugin-babel), though YMMV (I don't use it personally, so I'm not 100% sure it's kept up with changes to Babel) – in theory it will generate *slightly* leaner output, and the build will be faster than transpiling the whole bundle.