@jayphelps 2016-10-24
LGTM, we're going to continue to discuss more accurate source map support.