@zertosh 2016-10-24
@Rich-Harris: I tried `rollup-plugin-babel` but it didn't actually dedupe the helpers :/ The best output I got was from transpiling after the bundling. As far as build speed impact, this adds a second or two - but that's only really run for publishing, doesn't really affect dev-time speed.