@zertosh 2016-10-24
Thanks @jayphelps and @kwonoj! [Nuclide](https://github.com/facebook/nuclide) would really appreciate a `5.0.0-rc.2` release so we can start using this 😄