@jayphelps 2016-10-24
@zertosh side question, we've been trying to accumulate use cases from our Electron users who don't use a bundler. Since Nuclide is one of them, can you provide any insight? If a bundler is used, operator patching becomes the way to reduce the size of your shipped app.

We've heard of several who are worried about file size but don't use bundling to reduce this (and reduce `require()` file system pressure), but we don't have any real world experience with Electron to know if there are additional nuances that make bundling impractical, etc. If build times were the concern Electron is unique in the sense that you could get away with not bundling when doing local dev, but then bundle as part of a final packaging step.