@zertosh 2016-10-24
@jayphelps we don't bundle in Nuclide – for various reasons: we have several node binary modules, code in multiple languages (python and Java), and Nuclide is actually two apps the "client" and the "server" (we'd need two bundles - more dev overhead). All of these make bundling more complicated than what we're willing to deal with right now.

Our concern is not so much with size but load times - though there's some correlation between the two, that's not always the case. Since bundling is not option right now, the next best thing is to load the UMD bundle - which is considerably faster to load than the unbundled entry point. This is transparent to devs because at transpile time, we [convert](https://github.com/facebook/nuclide/blob/8afec4a8e1beb6c015522d9f705b57320fcf2092/pkg/nuclide-node-transpiler/lib/use-minified-libs-tr.js#L16) the source on `import`s from `rxjs` to `rxjs/bundles/Rx.min.js`.

Loading `rxjs` in Node is not cheap:

```
$ for i in {1..5}; do node -p "s=Date.now();require('rxjs');Date.now()-s"; done
178
187
185
184
191
```

I'd love it if the UMD bundle were brought down to the root of the package - so that it's easier to manually require in scripts. Maybe rename `rxjs/Rx.js` to `rxjs/Rx.all.js` (but still as the entry point), and `rxjs/bundles/Rx.js` to `rxjs/rx.js` (lowercase).