@kwonoj 2016-10-24
@jayphelps we don't bundle it as well due to similar reason to nuclide. In short, when it comes electron having various native-binary modules, it is not trivial effort to setup bundler (to exclude those modules, and setup other routing to load those modules as well) in addition to it requires 2 set of bundles eventually, one for main process and other for browser renderer process. 

What we do in our codebase for Rx is straightforward - instead of import everything, import patch operator individually for necessary ones. (import '.../add/../xxx')