@zertosh 2016-10-24
Oh, and we don't patch import operators because we don't have a way to enforce via [flow](https://flowtype.org) that you didn't forget something.