@kwonoj 2016-10-24
@zertosh yup, we're able to do it since we're using TypeScript in our codebase.