@kwonoj 2016-10-24
@zertosh but idea's simple - as latest typescript compiler (or language service) do understands module argumentation, if you write any consumer code like

```js
import {Observable} from 'rxjs/Observable';
import 'rxjs/add/observable/of';

Observable.of //type inference will work since language service accepts module augmentation
```

(I just wrote down code so it might not work via ctrl-v, but I assume you get idea)