@trxcllnt 2016-10-28
@zertosh @kwonoj @jayphelps it seems after this PR the files in `dist/global` aren't in UMD wrappers any more. Am I missing something, is this expected, or is this a regression?