@jayphelps 2016-10-28
@trxcllnt I'm seeing it here: https://unpkg.com/@reactivex/rxjs@5.0.0-rc.1/dist/global/Rx.js

Also one of my demo jsbins still works: http://jsbin.com/jexomi/edit?html,js,output

Can you clarify?

```js
(function (f) {
  if(typeof exports === "object" && typeof module !== "undefined") {
    module.exports = f()
  } else if(typeof define === "function" && define.amd) {
    define([], f)
  } else {
    var g;
    if(typeof window !== "undefined") {
      g = window
    } else if(typeof global !== "undefined") {
      g = global
    } else if(typeof self !== "undefined") {
      g = self
    } else {
      g = this
    }
    g.Rx = f()
  }
})
```