@trxcllnt 2016-10-28
@jayphelps was this PR included in the rc.1 build?