@jayphelps 2016-10-28
@trxcllnt nope! good catch!! I'm seeing the same thing locally. I'll check it out locally tomorrow. can you create a ticket to track?

related: I've been planning to bring up at our next meeting the idea of publishing canary builds of master to npm. The community could then run their CI against it to keep an eye out for breaking changes and also bug us if we break something unknowingly _before_ we cut a real release.