@jayphelps 2016-10-28
@zertosh any initial thoughts? you mentioned you had used it in your local builds without issue?