@Brooooooklyn 2016-10-28
I think `TS => ES5 code + ES 6 module ` is a better way to bundle