@jayphelps 2016-10-28
@Brooooooklyn we're talking about UMD bundles, which cannot have ES modules in them (would be a syntax error)