@Brooooooklyn 2016-10-28
@jayphelps rollup is able to bundle ES6 module to UMD without babel, and if your code is compiled to ES5 by TypeScript using `tsc entry.ts -m ES2015 --target ES5`, you could do the `TS => ES5 code + ES6 module=>use rollup => ES5 UMD code` without babel.

I've do such things for a long time in our project https://github.com/teambition/teambition-sdk/blob/master/tools/tasks/bundle.ts#L32

😉